Reimplemented generic file delete function

+++ core/src/util.c
@@ -2418,22 +2418,24 @@ void UtilGetGeofenceDirectoryPath(char *path, size_t pathLen) {
 }
 
 /*!
- * \brief UtilDeleteTrajectoryFile deletes the specified trajectory and deletes it
+ * \brief UtilDeleteTrajectoryFile deletes the specified trajectory
+ * \param name
+ * \param nameLen
  * \return returns 0 if the trajectory is now deleted. Non-zero values otherwise.
  */
 int UtilDeleteTrajectoryFile(const char *name, const size_t nameLen) {
 	char filePath[MAX_FILE_PATH] = { '\0' };
 	UtilGetTrajDirectoryPath(filePath, sizeof (filePath));
 
+	if (name == NULL) {
+		errno = EINVAL;
+		LogMessage(LOG_LEVEL_ERROR, "Attempt to call delete on null trajectory file");
+		return -1;
+	}
 	if (strstr(name, "..") != NULL || strstr(name, "/") != NULL) {
 		errno = EPERM;
 		LogMessage(LOG_LEVEL_ERROR,
 				   "Attempt to call delete on trajectory file and navigate out of directory");
-		return -1;
-	}
-	if (name == NULL) {
-		errno = EINVAL;
-		LogMessage(LOG_LEVEL_ERROR, "Attempt to call delete on null trajectory file");
 		return -1;
 	}
 	if (strlen(filePath) + nameLen > MAX_FILE_PATH) {
@@ -2450,21 +2452,23 @@ int UtilDeleteTrajectoryFile(const char *name, const size_t nameLen) {
 }
 
 /*!
- * \brief UtilDeleteGeofenceFile deletes the specified geofence and deletes it
+ * \brief UtilDeleteGeofenceFile deletes the specified geofence
+ * \param name
+ * \param nameLen
  * \return returns 0 if the geofence is now deleted. Non-zero values otherwise.
  */
 int UtilDeleteGeofenceFile(const char *name, const size_t nameLen) {
 	char filePath[MAX_FILE_PATH] = { '\0' };
 	UtilGetGeofenceDirectoryPath(filePath, sizeof (filePath));
 
-	if (strstr(name, "..") != NULL || strstr(name, "/") != NULL) {
-		errno = EPERM;
-		LogMessage(LOG_LEVEL_ERROR, "Attempt to call delete on geofence file and navigate out of directory");
-		return -1;
-	}
 	if (name == NULL) {
 		errno = EINVAL;
 		LogMessage(LOG_LEVEL_ERROR, "Attempt to call delete on null geofence file");
+		return -1;
+	}
+	if (strstr(name, "..") != NULL || strstr(name, "/") != NULL) {
+		errno = EPERM;
+		LogMessage(LOG_LEVEL_ERROR, "Attempt to call delete on geofence file and navigate out of directory");
 		return -1;
 	}
 	if (strlen(filePath) + nameLen > MAX_FILE_PATH) {
@@ -2477,6 +2481,40 @@ int UtilDeleteGeofenceFile(const char *name, const size_t nameLen) {
 		return -1;
 
 	strcat(filePath, name);
+	return deleteFile(filePath, sizeof (filePath));
+}
+
+/*!
+ * \brief UtilDeleteGeofenceFile deletes the specified file and deletes it
+ * \param pathRelativeToWorkspace
+ * \return returns 0 if the geofence is now deleted. Non-zero values otherwise.
+ */
+int UtilDeleteGenericFile(const char *pathRelativeToWorkspace, const size_t nameLen) {
+	char filePath[MAX_FILE_PATH] = { '\0' };
+	UtilGetTestDirectoryPath(filePath, sizeof (filePath));
+
+	if (pathRelativeToWorkspace == NULL) {
+		errno = EINVAL;
+		LogMessage(LOG_LEVEL_ERROR, "Attempt to call delete on null generic file");
+		return -1;
+	}
+
+	if (strstr(pathRelativeToWorkspace, "..") != NULL) {
+		errno = EPERM;
+		LogMessage(LOG_LEVEL_ERROR, "Attempt to call delete on generic file and navigate out of directory");
+		return -1;
+	}
+
+	if (strlen(filePath) + nameLen > MAX_FILE_PATH) {
+		errno = ENOBUFS;
+		LogMessage(LOG_LEVEL_ERROR, "Generic path name too long");
+		return -1;
+	}
+
+	if (filePath[0] == '\0')
+		return -1;
+
+	strcat(filePath, pathRelativeToWorkspace);
 	return deleteFile(filePath, sizeof (filePath));
 }
